Two newly created files (for adding a new job app)
"newApplication.php" presents a webform to the user through which they can create a new job application for their calendar.
"createApplication.php" is a server-side-only script, which receives the information submitted in the newApplication.php form, and updates the database. 
Note that both of these files use the "classes.php" file (uploaded moments before this upload).