Two newly created files for displaying user's details
"userDetails.php" displays the user's details (username, phone number, etc) and allows the user to edit any details.
"changeDetails.php" is a server-side-only script. If the user edits any of their details on the userDetails.php page, the changes are
sent to the changeDetails.php, which updates the database.
Note that both files require the "classes.php" file (uploaded moments ago).